* inline prop `tid` into derived classes to fix https://github.com/doctrine/orm/issues/4131 @ `Post` * move common props `postedAt` & `(dis)?agreeCount` from derived classes of `Post` into it @ `Post`
* mark abstract classes with attribute `MappedSuperclass`
@ `App\Entity`
@ symfony
