+++ symfony/src/Entity/Post/Content/PostContent.php
@@ -4,6 +4,7 @@ namespace App\Entity\Post\Content;
 
 use Doctrine\ORM\Mapping as ORM;
 
+#[ORM\MappedSuperclass]
 abstract class PostContent
 {
     #[ORM\Column] private ?string $protoBufBytes;
