+++ symfony/src/Entity/Post/Content/SubReplyContent.php
@@ -2,8 +2,10 @@
 
 namespace App\Entity\Post\Content;
 
+use App\Repository\Post\Content\SubReplyContentRepository;
 use Doctrine\ORM\Mapping as ORM;
 
+#[ORM\Entity(repositoryClass: SubReplyContentRepository::class)]
 class SubReplyContent extends PostContent
 {
     #[ORM\Column, ORM\Id] private int $spid;
