+++ symfony/src/Entity/Post/SubReply.php
@@ -8,11 +8,14 @@ use Doctrine\ORM\Mapping as ORM;
 #[ORM\Entity(repositoryClass: SubReplyRepository::class)]
 class SubReply extends Post
 {
+    #[ORM\Column] private int $tid;
     #[ORM\Column] private int $pid;
     #[ORM\Column, ORM\Id] private int $spid;
-    #[ORM\Column] private int $postedAt;
-    #[ORM\Column] private ?int $agreeCount;
-    #[ORM\Column] private ?int $disagreeCount;
+
+    public function getTid(): int
+    {
+        return $this->tid;
+    }
 
     public function getPid(): int
     {
@@ -22,20 +25,5 @@ class SubReply extends Post
     public function getSpid(): int
     {
         return $this->spid;
-    }
-
-    public function getPostedAt(): int
-    {
-        return $this->postedAt;
-    }
-
-    public function getAgreeCount(): ?int
-    {
-        return $this->agreeCount;
-    }
-
-    public function getDisagreeCount(): ?int
-    {
-        return $this->disagreeCount;
     }
 }
