+++ symfony/src/Entity/TimestampedEntity.php
@@ -4,6 +4,7 @@ namespace App\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
 
+#[ORM\MappedSuperclass]
 abstract class TimestampedEntity
 {
     #[ORM\Column] private int $createdAt;
